update accorion (use selenide)
